+++ apps/web/src/app/(protected)/clusters/page-view.tsx
@@ -33,9 +33,9 @@ import Link from 'next/link'
 import type { KubernetesCluster } from '@ror/js-api-client'
 import { User } from 'next-auth'
 import { CodeSnippet } from '@ror/react'
-import { ClusterCard, ClusterCardDisplayData } from '@/components/ui/cluster/cluster-card'
+import { ClusterCard } from '@/features/cluster/components/cluster-card'
 import { ClustersTable } from './cluster-table'
-import { ClusterSearch } from '@/components/ui/cluster/cluster-search'
+import { ClusterSearch } from '@/features/cluster/components/cluster-search'
 import { SortSelect } from '@/components/ui/sort-select'
 import { TabsViewSwitcher } from '@/components/ui/tabs-view-switcher'
 import { Toggle } from '@/components/shadcn/toggle'
@@ -52,6 +52,7 @@ import { cn } from '@/utils/clsxm'
 import type { WorkSheet } from 'xlsx'
 import { NotReadyMessage } from '@/components/ui/not-ready-message'
 import { loadMoreClusters } from '@/utils/cluster-actions'
+import { ClusterCardDisplayData } from '@/features/cluster/types/display-data'
 
 interface Params {
   view?: 'grid' | 'list'
